fix up bill charts, refactor layout

+++ dashboard/query/bills/bills_total_by_region.sql
@@ -1,6 +1,6 @@
 select 
     region,
-    sum("Spend_GBP_no_VAT"::numeric::money) as total_bills
+    sum("Spend_GBP_no_VAT"::numeric) as total_bills
 from 
     bills
 group by
